requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.3
psycopg2-binary==2.9.10
Pillow==11.2.1
drf-yasg==1.21.10
django-filter==25.1
django-celery-beat~=2.8.1
geopy~=2.4.1
celery~=5.5.3